Migrate verify invite endpoint with integration test
