--- server/service/endpoint_invites.go
@@ -1,1 +1,0 @@
-package service
